requirements.txt
langchain==0.3.27
chromadb==1.1.1
openai==2.2.0
streamlit==1.45.1
python-dotenv==1.1.0
langchain-openai==0.3.35
langchain-community==0.3.30
langchain-chroma==0.2.6
langchain-huggingface==0.3.1
sentence-transformers==5.1.1
transformers==4.57.0
torch==2.8.0